babel: extract declarations checker
